Added frame time functions

+++ nl_lib/nl_memory.h
@@ -1,7 +1,7 @@
 #ifndef __NL_MEMORY_H__
 #define __NL_MEMORY_H__
 
-void *memory_allocate(unsigned long int size);
+void *memory_allocate(unsigned long long size);
 void memory_free(void* memory);
 
 #endif //__NL_MEMORY_H__
